Initial Upload - V1.4.5
Initial upload of the Foothold mission files. To use remember to make a new mission in DCS. Open with WinRar/7z and drag all files into that .miz file. Do not unpack the .miz unless you are editing and plan to put the files back. This will mess with DCS compression and the software will not recognize it.